feat: dynamic id for components with mandatory id feat: options attribute made optional in the select component fix: rating component
