+++ projects/design-angular-kit/src/lib/abstracts/abstract.component.ts
@@ -4,6 +4,7 @@ import {
   Component,
   ElementRef,
   EventEmitter,
+  inject,
   Input,
   OnChanges,
   Output,
@@ -12,23 +13,34 @@ import {
 } from '@angular/core';
 
 @Component({ template: '' })
-export class AbstractComponent implements AfterViewInit, OnChanges {
+export abstract class AbstractComponent implements AfterViewInit, OnChanges {
 
   /**
    * The element ID
    */
-  @Input() id?: string;
+  @Input() id: string = this.getDefaultId();
 
   /**
    * Fired when component input attributes was changed
    */
   @Output() public valueChanges: EventEmitter<void>;
 
-  constructor(
-    protected readonly _renderer: Renderer2,
-    protected readonly _elementRef: ElementRef,
-    protected readonly _changeDetectorRef: ChangeDetectorRef
-  ) {
+  /**
+   * Counter of active instances
+   * @private
+   */
+  private static instances = 0;
+
+
+  protected readonly _renderer: Renderer2; // Injected
+  protected readonly _elementRef: ElementRef; // Injected
+  protected readonly _changeDetectorRef: ChangeDetectorRef; // Injected
+
+  constructor() {
+    this._renderer = inject(Renderer2);
+    this._elementRef = inject(ElementRef);
+    this._changeDetectorRef = inject(ChangeDetectorRef);
+
     this.valueChanges = new EventEmitter<void>();
   }
 
@@ -38,5 +50,15 @@ export class AbstractComponent implements AfterViewInit, OnChanges {
 
   ngOnChanges(changes: SimpleChanges): void {
     this.valueChanges.next(); // The inputs were changed
+  }
+
+  /**
+   * Generate unique id for components
+   * @private
+   */
+  private getDefaultId(): string {
+    const name = this.constructor.name.replace('Component', '');
+    const kebabName = name.replace(/[A-Z]+(?![a-z])|[A-Z]/g, ($, ofs) => (ofs ? '-' : '') + $.toLowerCase());
+    return `it-${kebabName}-${AbstractComponent.instances++}`;
   }
 }
