+++ projects/design-angular-kit/src/lib/components/core/progress-bar/progress-bar.component.ts
@@ -5,7 +5,6 @@ import { BooleanInput, isTrueBooleanInput } from '../../../utils/boolean-input';
 @Component({
   selector: 'it-progress-bar[value]',
   templateUrl: './progress-bar.component.html',
-  styleUrls: ['./progress-bar.component.scss'],
   changeDetection: ChangeDetectionStrategy.OnPush
 })
 export class ProgressBarComponent {
