+++ projects/design-angular-kit/src/lib/components/utils/language-switcher/language-switcher.component.html
@@ -1,4 +1,4 @@
-<it-dropdown id="language-switcher">
+<it-dropdown>
   <ng-container button>
     <span class="visually-hidden">{{'it.utils.selected' | translate: { lang: (currentLang$ | async)?.label } }}</span>
     <span>{{(currentLang$ | async)?.label || ('it.utils.select-language' | translate)}}</span>
